cleaned up code

+++ src/utils.py
@@ -218,39 +218,6 @@ def get_zero_pt(
             (-alpha * 2**bit_width) / (beta - alpha)
         )
 
-# def run_inference(
-#     model: nn.Module,
-#     loader: DataLoader,
-#     n_graphs: int,
-#     n_trivial_preds: int,
-#     device: torch.device = torch.device("cpu")
-# ) -> float:
-#     sigmoid = nn.Sigmoid()
-#     correct_preds = 0
-
-#     # loop over batches
-#     with torch.no_grad():
-#         for batch in tqdm(loader):
-#             # unzip data
-#             x = batch.x.to(device)
-#             edge_index = batch.edge_index.to(device)
-#             edge_attr = batch.edge_attr.to(device)
-#             batch_label = batch.batch.to(device)
-
-#             out = model(
-#                 x,
-#                 edge_index,
-#                 edge_attr,
-#                 batch_label,
-#             )
-
-#             prediction = (sigmoid(out.detach()) > 0.5).long()
-#             target = batch.y.int()
-#             correct_preds += int((prediction == target).sum())
-
-#     accuracy = (n_graphs - correct_preds - n_trivial_preds) / n_graphs
-#     return accuracy, correct_preds
-
 def run_inference(
     model: nn.Module,
     loader: DataLoader,
@@ -277,6 +244,6 @@ def run_inference(
             )
 
             prediction = (sigmoid(out.detach()) > 0.5).long()
-            target = batch.y.to(device).int()
+            target = batch.y.to(device).long()
             correct_preds += int((prediction == target).sum())
     return correct_preds
